// src/api/generated/iMSIncidentManagementSystem.schemas.ts
export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  guard: 'guard',
  operator: 'operator',
  analyst: 'analyst',
  admin: 'admin',
} as const;
// ...
export interface UserCreate {
  email: string;
  first_name: string;
  last_name: string;
  surname?: string | null;
  role: UserRole;
  password: string;
}

export interface UserUpdate {
  email?: string | null;
  first_name?: string | null;
  last_name?: string | null;
  surname?: string | null;
  role?: UserRole | null;
  is_active?: boolean | null;
}